Sistem Inventori Perangkat TI Kantor Pos Cikarang/requirements.txt
alabaster==0.7.9
alembic==0.8.10
aniso8601==1.2.0
appdirs==1.4.0
argh==0.26.2
Babel==2.3.4
blinker==1.4
click==6.7
coverage==4.3.4
Flask==0.12.2
Flask-Admin==1.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.2
fpdf==1.7.2
gunicorn==20.0.4
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
mysqlclient==2.0.2
pdfkit==0.6.1
petl==1.1.1
protobuf==3.14.0
pycparser==2.20
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
pywebview==3.4
six==1.10.0
SQLAlchemy==1.1.11
WTForms==2.1
urllib3==1.25.9
Werkzeug==1.0.1
uwsgi
